Optional: Refactor InternalToString
