+++ src/core-taggeds-optional/Optional/InternalToString/InternalToString.InnerAbsent.cs
@@ -1,0 +1,9 @@
+﻿namespace System;
+
+partial class InternalToString<T>
+{
+    private static class InnerAbsent
+    {
+        public const string ValueString = "()";
+    }
+}
