--- src/core-taggeds-optional/Optional/InternalToString/InternalToString.InnerConsts.cs
@@ -1,9 +1,0 @@
-﻿namespace System;
-
-partial class InternalToString<T>
-{
-    private static class InnerConsts
-    {
-        public const string Absent = "()";
-    }
-}
